feacture/py-views: Sedividieron las views en archivos diferentes.
