+++ backend_vibe/app_vibe/urls.py
@@ -1,24 +1,26 @@
 from django.urls import path
-# from rest_framework import routers
-
-# from .views import RegistroView, LoginView, LogoutView, profileView, SubirArchivoApiView, mi_vista,user_detail, user_config
-from app_vibe import views
-# routers=routers.DefaultRouter()
-# routers.register(r"User", UserViewSet)
+from app_vibe.view import auth_views, user_views, profile_views, conexion_views
 
 urlpatterns = [
-    # path('', include(routers.urls)),
-
-    path('register/', views.RegistroView, name='register'),
-    path('login/', views.LoginView, name='login'),
-    path('logout/', views.LogoutView, name='logout'),
-    path('profile/', views.profileView, name='logout'),
-    path('Subir/', views.SubirArchivoApiView.as_view(), name='Subir'),
-    path("tem/", views.mi_vista, name="tem"),
-    path('user/<int:user_id>/', views.user_detail, name='user-detail'),
-    path("user_config/<int:user_id>/", views.user_config, name="user-config")
-    path('Publications/', PublicacionView, name='Publicacion'),
-    path("Publications/create/", PublicacionCreateView, name="Publicacion-create"),
-    path("Comments/", CommentarioView, name="Comentario"),
-    path("Comments/create/", CommentarioCreateView, name="Comentario-create"),
+    # auths
+    path('auth/register/', auth_views.RegistroView, name='auth-register'),
+    path('auth/login/', auth_views.LoginView, name='auth-login'),
+    path('auth/logout/', auth_views.LogoutView, name='auth-logout'),
+    
+    # user
+    path('users/<int:user_id>/', user_views.user_detail, name='user-detail'),
+    path("users/<int:user_id>/config/", user_views.user_config, name="user-config"),
+    
+    # profile
+    path('profile/', profile_views.profileView, name='logout'),
+    
+    # publicactions
+    path('Publications/', conexion_views.PublicacionView, name='Publicacion'),
+    path("Publications/create/", conexion_views.PublicacionCreateView, name="Publicacion-create"),
+    
+    #coments
+    path("Comments/", conexion_views.CommentarioView, name="Comentario"),
+    path("Comments/create/", conexion_views.CommentarioCreateView, name="Comentario-create"),
+    
+    path('upload/', conexion_views.SubirArchivoApiView.as_view(), name='Subir'),
 ]
